Remove serving of UI static files
Previously this package was able to serve the HTML, JS and CSS of
Stencila user interfaces. This allowed a user to download this package
and use Stencila directly, through their browser, without having to
download Stencila Desktop as well. While we may reintroduce that
later, to reduce the scope of this package, it is currently being removed